+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bug fix, EPSS to use current UTC time 
+- Bug fix, validation command has a seperate bundle function to prevent error overwriting on recursive calls
+
+### Added
+
+- Lazy File Reader in internal/io to open file errors at read
 
 ## [0.1.0] - 2023-07-26